Arb clk DCM version update
The pcore used for the arb clk DCM has changed from d to e in the later
versions of EDK so the module instantiation in the mhs file was fixed.

This means using the arb clk on the 11.5 tools wont work. Just change d back
to e if you really still use 11.5.