Add zdok_rev parameter to ADC16 yellow block
A second differential input ADC16 board was created that uses fewer clocks from
the ADCs and programs the ADCs via ZDOK pins.  This new board therefore has a
different ZDOK pinout than the original board.

The original board's ZDOK pinout is known as "ZDOK Pinout Revision 1".
The new board's ZDOK pinout is known as "ZDOK Pinout Revision 2"

A singled-ended variant of the ADC16 board also uses "ZDOK Pinout Revision 2".

This commit only adds the parameter to the yellow block, but it is not yet used
to define the pinout.  That change will be in a future commit.