Make reuse_block more likely to reuse block
reuse_block used to compare the source block with the lower-cased name
reference block.  This change now uses strcmpi to perform the comparison, which
ignores the case of both the source block and the reference block.  This leads
to more blocks being reused and fewer blocks being unnecessarily replaced due
to case differences.

Log messages from reuse_block have also been improved to include the source
type and, if different, the reference block type in the message.