Use "built-in" blocks whenever possible
Some blocks can be added as "simulink/Foo/Bar" or as "built-in/Bar".  Using the
latter form, when possible, makes reuse_block work better.  Passing
"simulink/Foo/Bar" to reuse_block when the block can (and will) be added via
"built-in/Bar" will always cause the block to be replaced rather than reused as
desired.