mdl2m script
mdl2m allows the generation of a script that will create and populate a subsystem
to replicate the one provided. It currently iterates through all component blocks
and creates functions that will generate them in an empty system.

The intention is to add functionality to mdl2m.m so that it can be run against
an arbitrary subsystem, allowing scripts to be generated for all library
blocks or application systems and make it easier to work with these in
revision control systems i.e store the script instead of the binary-like
mdl or slx file.

xBlocks has a similar function but this only allows Xilinx systems to
be generated, if hdl coder or other third party tools are used we will need
to have a vendor agnostic tool to do this.

TODO: generate built-in blocks
TODO: Xilinx blocks
TODO: add lines between blocks
TODO: intelligent sort algorithm