Bus section added to casper_library
Contains blocks that operate on busses consisting of multiple
parallel data streams. Useful when the same operation must be performed on
an arbitrary number of streams (e.g FIR) with shared control logic.

The Bus library is generated with a script (casper_library_bus_init.m) when
casper_library is loaded. casper_library_init  was generated using the mdl2m
function and any changes to the library, when saved, cause mdl2m to regenerate
this script which can be committed to a revision control system. This should
ensure that changes to the Bus library are easy to see (diff the new script
against the old) and should make merge conflicts easier to resolve.