Handle newlines in source name in reuse_block
Reuse block compares the name of the existing block's reference block (aka
"source") to the requested block type.  The requested block type sometimes
contains space characters (e.g. "simulink/Logic and Bit Operations/Compare to
Constant").  The actual block name in the library uses newlines in place of
some of these spaces, so such requests never match so the block was always
replaced.

This commit replaces newlines in the source block name with spaces so that the
match will succeed and such blocks will be properly reused.