Fix lazy initialization in dump_exception
The comparison of the MException object being dumps with the last_dumped
variable caused an error with last_dumped being uninitialized so the block
inside the conditional never ran so last_dumped never got initialized.  This
fixes it be running the block if last_dumped is not an MException object (e.g.
if it is uninitialized) or not the same MException object that is being dumped.