Fanout control in bus_mult
The bus_replicate block now has the capability of adding pipeline
blocks with configurable latency to every replicated bus.

The bus_mult_init script now adds a replication layer consisting
of bus_replicate blocks just after input. The script calculates
the fanout for the inputs and replicates the inputs appropriately
to reduce this fanout to the maximum specified by the user.