Cache bram_block_custom netlists
The bram_block_custom pcore uses coregen to create BRAM netlists for "Shared
BRAM" blocks based on the options specified by the user.  This provides much
greater flexibility such as the ability to pipeline the BRAM primitives and
overall core.  Unfortunately, running coregen is somewhat slow (I think it has
a long startup time) so models that use lots of BRAMs take an exceedingly long
time to build (not even counting the map and par steps) due to running coregen
for each Shared BRAM.  This can be especially frustrating if the BRAMs are all
the same since mutiple runs of coregen are being used to create equivalent
netlists over and over.

This commit makes bram_block_custom's BRAM generation smarter by caching
generated netlists.  If a cached netlist exists for the given options, it is
used instead of re-running coregen to generate an equivalent netlist.  If a
cached netlist does not exist, coregen is run as usual and the resulting
netlist is cached for use by subsequent BRAMs having the same options.

The cache directory is named "cache_dir" and it lives next to the
XPS_ROACH2_base directory.  This keeps it unique to each design yet persists
over multiple builds to maximize the benefit of caching.  Currently the only
way to share the netlist cache across multiple designs is to symlink
"cache_dir" to a common directory.  It is not possible at present to specify
the location (or name) of "cache_dir" via an environment variable.

Here is an outline of the new netlist generation process:

1) Build the text string for input file to coregen.

2) Compute SHA1 hash of that text string.

3) If a "<sha1>.ngc" file exists in cache_dir, copy it to output directory (and
   rename it to "bram.ngc").

4) If the cached netlist does not exist, run coregen and save the "bram.ngc"
   netlist file to "<cache_dir>/<sha1>.ngc" for future use.