Handle missing parameters better in xps_bram
The new Shared BRAM mask has three new parameters that may not be present in
existing models with inactive links to the Shared BRAM block in the XPS
library.  When creating an xps_bram object for such a block, we need to be able
to cope with such blocks rather than cause an exception.  This commit does that
by assuming backwards-compatible values for these parameters then ignores
errors when trying to get their real values from the block.

If the Shared BRAM block in the model does not have the first new parameter, a
warning is issued to alert the user that the block's link needs to tbe
restored.