Make Shared BRAM parameters backward compatible
Two of the new shared BRAM parameters affect the latency of the Shared BRAM
block.  This change defaults these parameters to values that provide the same
latency as the version without these new parameters.  This means that a user
can restore the Shared BRAM links of existing models to get the new mask (with
the new parameters) withOUT affecting the behavior of the model.  Then the user
can update the new options and deal with change in latency at their convenience.