Use munge_block in adc16_init
This diables the link to the ADC16 library block in order to preserve the
custom drawn state.  Without this, opening a dual ADC16 model would show a
single ADC16 yellow block even though the mask said to use 2 boards.  This is
bacause the internal state of the ADC16 block was restored from the library
(since the link was still active) and the init script did not redraw it to
match the mask parameters because the "same_state" function returned true
(same_state only compares a hash of the mask parameters with the value stored
in the UserData parameter).