Update uniform-to-normal block w/ DDS 4.0 compiler
The uniform-to-normal (u2n) block used a SineCosine block to generate the sine
and cosine values used for the Box-Muller transform.  The SineCosine block has
been discontinued and replaced by the "DDS 4.0 Compiler" block.

This commit changes the u2n block in casper_library_sources.mdl to use the "DDS
4.0 Compiler" block instead of the SineCosine block.  While still performing
the same transform, the outputs of the new u2n block will not be identical to
the output of the old u2n block since 1) the DDS 4.0 Compiler block's output is
not the same as the SineCosine block and 2) the latency of the DDS 4.0 Compiler
block is longer than the SineCosine block.

Systems that correlate the outputs of an old u2n block with a new u2n block
will notice these differences as slightly reduced correlation and a non-zero
relative delay.

The wgn_test.mdl test file and the reference output file also updated.