Make reuse_block more pedantic
For some unknown reason, find_system will sometimes return more than one block
for a given name.  Sometimes it's simply just more than one reference to the
same block, which is weird but OK.  Othertimes it really thinks there are two
different blocks with the same name, which is indicative of some bizarre error
condition that needs attention.  This change makes reuse_block a bit more
pedantic about how it handles multiple blocks being returned by find_system.
If the multiple blocks are simply duplication of the same block, it will
proceed normally, but if the multiple blocks are distinct, it will throw an
exception.