downconverter library generation, cosin block
The downconverter section of the library is now generated (using
casper_library_downconvert_init.m) as the bus section is. Any saved
changes will trigger the re-generation of this script (via mdl2m.m).
This should help make merge conflicts less likely and easier to solve.

The cosin block has been added to the Downconverter library. This allows
complex exponentials to be generated with various optimisations. Will be
used in future DDC and FFT improvements.

casper_library.mdl is in Simulink version 8 format, will attempt a back-port
to 7.9 in a future commit.