Fix auto_tap and baseline_tap mask initializers
The mask inializers for the auto_tap and baseline_tap blocks did not pass all
of the mask parameters to the initialization function.  This caused save_state
to inaccurately reflect the state of the block, which could erroneously cause
same_state to prevent updating the blocks' internals.  It also caused a problem
for the backpopulate_masks funtion ("No value specified for n_ants").

These blocks now use the recently added get_mask_params function to fix the
above problems.