update_casper_block now stages new block in root
update_casper_block used to stage the new block in the same subsystem as the
old block.  This sometimes had the unintended consequence of running the
subsystem's mask init script.  Some such scripts call "clean_blocks" to delete
blocks with dangling inputs such as the new block being staged.

To fix this, update_casper_block now stages the new block in the root of the
model using the name "__x__tmp__x__".  This will fail if there is already a
block at the root level with that name, but that risk is negligible.  After
staging the new block there, the old block is deleted and the new block is
copied into its place and the temporary staging block is deleted.