Stop qdr block from whining when in CASPER library.
The QDR mask initializer looks for an XSG block to determine the hardware
platform.  If it can't find one it pops up a warning dialog and displays a
warning message.  This is OK if the block is in a user model, but when it is in
a CASPER library this condition is expected so the whining is not needed.