Make update_casper_block not use temporary block
update_casper_block used to add a temporary new block, copy the parameters of
the old block, delete the old block, then move the new block into the old
block's place.  The new approach builds a cell array of mask names and values
to copy to the new block, then deletes the old block, and adds the new block as
the old block using a special positioning sequence to control port
connectivity.  This probably reverts back to the problem of firing the mask
init script of the parent block which might try to run clean_blocks.