Use forwarding table in update_casper_block
casper_library's forwarding table maps old (now very old) block names to new
block names.  It is used automatically with add_block, but not with get_param.
Because the source block name that update_casper_block passes to get_param may
be very old, update_casper_block must explicitly map it to the new name, if
applicable.

To accomplish this, the casper_library_forwarding_table.m function has been
updated.  If given no parameters, it still returns a cell array suitable for
storing as casper_library's ForwardingTable.  If it is given a block name, it
returns the current name for that block name.  If the name is an old name in
the forwarding table, the new name will be returned; otherwise the name passed
in will be returned unchanged.