Coeff gen change
Changes made to coeff_gen to use the cosin block to reduce BRAM usage. This
has caused changes to ripple through other blocks around it.

FFT seems to work (simulates ok for simple test cases). fft_wideband_real does not but
this seems to be the same as versions pre these changes. Compiles up to
13 stages work but above this causes an exception (to be investigated).

TODO:
* support some old FFT options (multiplier implementation etc)
* fix fft_wideband_real
* improve performace (2^12 FFT capable of 236MHz before, now 213MHz)
* add asycnronous operation option to all fft blocks
* save more BRAM by making biplex_fft handle multiple input streams so
that coefficients can be shared.
* generate coefficients in fft_direct by having a starting vector of (cos0, -sin0)
and rotate this vector by n/2pi (as noted by Ryan in a recent email). This will
remove the largest coefficient lookup tables.