Fix core_info.tab offsets and sizes for ROACH1 QDR
In commit 1cfc60ea, the QDR's system.mhs generation script changed the base
address of the CPU interface for QDR1 on both ROACH and ROACH2.  The
core_info.tab template for ROACH2 was changed accordingly, but the
core_info.tab template for ROACH1 was not so lucky.

This commit changes the ROACH1 core_info.tab file so that the offsets and sizes
for qdr0_memory and qdr1_memory on ROACH1 match the addresses of the
qdr0_sniffer and qdr1_sniffer instances in system.mhs.