Add ability to specify custom hw_routes file
Users now have the option of loading a custom hw_routes file by setting the
environment variable CASPER_HW_ROUTES to the name of the file to load.  If not
set, the default file "hw_routes.mat" is loaded.

The file name given is passed to the Matlab "load" function.  It can be given
as a fully qualified file name (i.e. start with a '/') or a partially qualified
file name.  Partially qualified names (including file names with no '/'
characters) are searched for on the Matlab path.

One use of this feature is to specify an alternate hw_routes file for ROACH2
Revsion 1 development.