Update adc5g_init's handling of Downsample block
The Downsample block in newer versions of the DSP toolbox have two new
parameters that must be set as shown below for proper handling of the sim
inputs of some of the ADC blocks.  Older versions of the DSP toolbox do not
have these options, so they must be set in a separate try/catch block to ensure
backwards compatiblity.  The two new parameters and the values that must be set
to are:

'InputProcessing' = 'Elements as channels (sample based)'
'RateOptions'     = 'Allow multirate processing'