Add block name to adc16's state info
The adc16_init script uses the same_state/save_state mechanism to prevent
unnecessary redrawing of the block's contents.  When moving an ADC16 block into
a subsystem, the block's contents must be redrawn because the Gateway In names
need to be updated to reflect the block's new place in the design hierarchy (to
match the naming conventions expected/required by the CASPER toolflow).  To
force re-drawing after moving the ADC16 block into a subsystem (or after
renaming a parent subsystem), the block's name is now part of the block's state
that is saved with save_state and checked with same_state.