[adc5g] Set MMCM to explicitly use HIGH bandwidth mode.
Also started using the CLKOUT0 output which has 1/8 resolution in its
divide factor. This should allow better fine-tuning of the MMCM parameters
to try and meet the requirements of HIGH bandwidth mode. Also edited the
xps_adc5g.m script to handle these changes.

Also set the minimum PFD frequency in the xps_adc5g.m script to match the
actual minimum of the HIGH bandwidth mode case. Unfortunately this means that
there are no longer valid MMCM parameters that meet our requirements for our
highest sample rates, 5 GHz and 4.576 GHz. A fix for this is forth-coming;
Using LOW bandwidth mode is possible but not ideal.