[adc5g] Added a BUFR divide to the ADC's clock input path before the MMCM.
This can divide the clock going into the MMCM which makes meeting the HIGH
bandwidth mode easier given all the limitations in parameters. This will only
work up to an ADC sample rate of 4800 MSps. It's unclear how to proceed beyond
that without tricking the tools into accepting the HIGH bandwidth mode for a
PFD frequency below 135 MHz.