Make same_state/save_state warn on empty params
If a block's mask is configured to evaluate a given edit field, accessing the
mask variable in the mask init pane will return the value that the edit field
string evaluated to.  Unfortunately, if evaluating the string raises an error,
the mask init callback will (or at least sometimes does) ignore the error and
pass an empty array for that parameter's value.  This can result in the mask
init script silently doing unexpected/wrong things.

This gets around that problem by logging and rethrowing an exception.  The
exception is logged because otherwise it will silently abort the mask init
script.  This way it non-silently aborts the mask init script.