Change how same_state/save_state throws exception
The initial implementation used dump_and_rethrow to ensure that the exception
was dumped before throwing it.  Unfortunately, dump_and_rethrow uses rethrow
which only work if the exception already has been thrown and caught.  This is
not the case in same_state/save_state, so now they just call dump_exception and
then throw.