Improve how update_casper_block uses add_block
Now update_casper_block uses add_block to place the new block directly in its
final subsystem rather than placing it at the top level.  Placing the block at
the top level did prevent the parent's mask init script from running, but it
ran the risk of running the block's mask init script without access to parent's
mask variables that the block's mask might reference.  The new block really
needs to be added directly to its subsystem instead of adding it to the top
level and then moving to its subsystem.  To prevent the parent's mask init
script from running, we save the value of parent's MaskInitialization
parameter, set it to an empty string, add the block, then restore the parent's
MaskInitialization parameter.

The newly placed block is also nudged around a little in position to make sure
it connects to the lines that were left behind when the old block was deleted.
It also bulges the block out and back slightly to get the port spacing correct.