Remove save_state's call to backpopulate_mask
It is unclear to me why backpopulate_mask is (still) called from save_state.  I
think the backpopulating of masks was started long ago when the evaluation of
mask parameters was buggy leading to lots of crashes.

In fact, calling backpopulate_mask in save_state can defeat user masks.  For
example, assume a user creates a masked subsystem with a mask parameter named
"DelayLen" that is currently set to 100.  The user then adds a CASPER block to
that subsystem and sets one of the CASPER block's mask parameters to
"DelayLen".  If the CASPER block's mask init script calls save_state which
calls backpopulate_mask, that parameter of the CASPER block will be
backpopulated to "100" based on the current value of the DelayLen parameter in
the user's mask.  If the user then changes their block's DelayLen parameter,
the CASPER block will not change to reflect the new value since its mask
parameter is now hardcoded to 100!

I think things are more stable now, so it's probably worth trying to live
without backpopulating masks in save_state.