Add ability to reset dump_exception
dump_exception stores the most recently dumped exception so that it will not
dump the same exception multiple times in a row as it bubbles up the call
stack.  Sometimes an exception being dumped happens to match the previously but
not recently dumped exception.  In that case, the exception is not dumped
thereby depriving the user of (potentially) useful information.  This commit
modifies dump_exception so that calling it with an empty value (i.e. '' or [])
will cause it to clear its stored exception.  This provides a way for scripts
to ensure that dump_exception will always dump the next exception being dumped.