Error out of FFT init scripts if block is too old
The new FFT init scripts rely on new mask parameters.  If these new mask
parameters are not present in the mask of an existing FFT block (i.e. one that
was originally placed from an earlier library version), then the new init
script will encounter all sorts of problems and leve the FFT block in a wholly
unusable state.  To prevent this problem, the FFT init scripts now verify that
the FFT block being initialized has one of these new parameters (n_streams) and
throws an exception alerting the user that the block is out of date if it
doesn't have that parameter.