Make same_state smarter when checking empty values
Empty parameter values passed to same_state can be indicative of problems.
same_state now vets empty parameter values and raises an exception if any
problems are found.  See comments in same_state.m for more details.

The save_state function has the old version of this functionality removed.
It is called too late to be a useful place to check for empty parameters and it
is redudant to check again after same_state has already done so..

The eval_param function's final eval attempt is now done outside of a try
block, so it can now throw an exception.  same_state now depends on this new
functionality of eval_param.

Now eval_param will only evaluate the mask parameter's string if it is
non-empty and is marked for evaluation in the mask dialog, otherwise it will
return the mask parameter string as-is.