Make coeff_gen_init get FPGA from SysGen block
Instead of determining FPGA characterisics based on the "hw_sys" parameter of
the CASPER XSG block, coeff_gen_init now gets it from the FPGA part parameter
of the System Generator block, which is more generally available.  For example,
a model used to create netlists for black boxing will have a System Generator
block, but it is not likely to have an XSG block.