Remove some special cases from update_casper_block
Commit c0505f3 added special handling to update_casper_block for some new FFT
parameters whose library values were not backwards compatible with older
versions of these blocks that lacked these new parameters.  As of commit
004cc6e those special cases are no longer needed in update_casper_block so they
are being removed.

The old block's UserData is copied to the new block if it looks like it was set
by save_state.  The "state" field is first cleared to ensure that the block
will be re-initalized by the mask init script.  The parameters field is left
intact in case some mask strings refer to it using the inelegant
"getfield(getfield(get_param(...)))" syntax.