Make new FFT drawing more backwards compatible
The new style FFT blocks have no contents when in the library.  When in a
library or when placed into a user's model, the mask init script will only draw
the contents of the block once the user has entered valid parameters.  Some of
the FFT blocks used a new parameter, n_streams, for this purpose.  When
n_streams was 0, the mask init script would not draw the block's contents.  The
n_streams value in these blocks' masks was 0 in the library so that the blocks
would not be drawn in the library or when the user first places the block into
a model.  This works as expected for user's creating new models, but it is
problematic for users trying to update existing FFT blocks do not have
n_streams using update_casper_block (or update_casper_blocks).

The update_casper_block function copies existing parameters values from the old
block to the new block.  For new parameters that do not exist in the old block,
update_casper_block let the new block use the default value from the library
block, but for the blocks described above, the default value from the library
for the new n_streams parameter was the "do nothing" value of 0.  Commit
c0505f3 addresses this problem in a rather non-sustainable way by recognizing
"n_streams" as a special case parameter and setting it to 1 if the old block
did not have an N-streams parameter.  It also had to special case the new
"bin_pt_in" parameter.

The current commit changes the new FFT drawing technique slightly for these
"n_streams" blocks.  The mask init scripts for these blocks no longer use
n_streams to determine whether to draw their contents.  They now do nothing if
the block lives in a library or if the (very old) parameter "FFTSize" is 0.
The blocks have been updated in the library to have n_streams set to 1 in their
masks.  This means that the n_streams parameter no longer needs to be special
cased in update_casper_block.  The FFTSize has been changed to 0 for these
blocks so that the user will still have a "do nothing" block when added to
their model until they enter in valid parameters (i.e. FFTSize > 0) in the mask
dialog.

Similar (though slightly different) handling was added to these blocks for the
bin_pt_in parameter.  That parameter is now set to -1 in these blocks' masks in
the library.  The init scripts recognize the -1 value to mean "num_bits_in-1",
which is backwards compatible with old models whose FFT blocks do not have a
"bin_pt_in" parameter.  This means that the bin_pt_in parameter no longer needs
to be special cased in update_casper_block.

This new approach also removes the init script code that would delete the
block's contents if the "do nothing" (fka "leave nothing") value was detected.
That seemed like a convenience for library maintainers, but it's easy enough to
delete the contents manually if desired.

This commit only applies these new drawing style changes to the blocks that
used n_streams==0 to determine when to delete contents.  Other blocks may be
content-free in the library and decide to draw or delete their contents based
on other parameters besides n_streams.  These blocks could likely benefit from
the same sort of changes being introduced here even if it is not necessary to
improve their backwards compatibility.

The changes to casper_library_ffts.mdl were made surgically with a text editor,
so the diff might actually be usable! :-)