More FFT backwards compatibility updates
Treat bin_pt_in==-1 (or bin_pt==-1) as special case in mask init scripts for
high level FFT blocks that were not modified earlier in commit 004cc6e.  See
that earlier commit for more details.  Also changed value to -1 for this mask
parameter in the blocks' masks in casper_library_ffts.mdl.

The changes to casper_library_ffts.mdl were made surgically with a text editor,
so the diff might actually be usable! :-)