Be more selective about evaluating in eval_param
eval_param now checks whether the parameter is an edit field before evaluating
the parameter.  This is in addition to the already existing checks for
non-emptyness of the parameter's string and the parameter's "evaluate" flag.

Checking for whether the parameter's "mask style" is "edit" is needed because
some masks have checkbox and/or popup parameters set to evaluate.  Because
these parameters have non-empty strings (e.g. "on" or "off" in the case of
checkboxes) and are set to evaulate, eval_param would try to evaluate their
value, which would rarely succeed and even more rarely (i.e. never) be what is
desired.