Added state machine debug probes to the QDR's "sniffer" OPB slave.
These allow checking the state of various state machines within the OPB
controller, specifically the PHY, the bit align and bit training logic.
This can be useful for figuring out if one of these are stuck in a certain
state which should narrow down problems, for example if sys_clk2x is not
present then bit_train will be stuck at idle.

To check the states read second word of the qdr_ctrl interface. For
example to read the states of QDR0 using Python/corr issue the following
command:

>>> print hex(roach2_10.read_uint('qdr0_ctrl', offset=2))
0xff000130

The two most significant nibbles will be 0xff to show support for this feature.
The three least significant nibbles will be the PHY state, the bit align state,
and the bit train state. In the above example the PHY is stuck in STATE_BIT_ALIGN,
the bit align logic is stuck in STATE_TRAIN, and the bit train logic is stuck in
STATE_IDLE. This indicated that the bit training logic was not clocking.