If enabled, the QDR is reset until the fabric and sys_clk MMCM's are locked.
This feature can currently be enabled by setting the QDR_MMCM_LOCK_EN environment
variable to something. It is not on by default because not all ADC's presently
break-out their MMCM locked signals. Eventually this feature should be enabled
via a checkbox in the QDR's yellow block mask.

Also in this commit I've broken the aux_clk MMCM locked signal in case it
is used as the fabric clock.