Added a "hard" reset line to QDR config.
This reset holds QDR reset high until it is de-asserted again by the OPB.
A normal "soft" reset can still be done by writing a 1 to offset 0 but if a
"hard" reset is needed this can be done by writing 256 to offset 0 instead.
Just be sure to write a 0 after that.