[opb_dram_sniffer] Removed logic for setting Sl_errAck in mem_opb_attach.v
I believe this was causing the PowerPC to hang when writing to or reading from
other OPB slaves since the error was being set when OPB_select=1 and the OPB_ABus
was out of range of the slave. This was not strictly needed so it's been removed
and it seems to fix the hanging issue.