bus_dual_port_ram added to bus
Allows a bus to be spread across multiple BRAMs and allows
the fanout of address, enable, write enable etc to be controlled.

Needed for pfb_fir coefficient storage and delays where large
busses are being stored in BRAMs with non-trivial address, write enable
etc logic.

Gives the same interface as a Xilinx dual port ram block. Parameters
are the same except that depth is inferred from the length of the
initialisation vector and bit width and binary point of the input
vectors must be specified.