[adc5g] Changes to reset logic.
The MMCM is now reset using dcm_reset instead of ctrl_reset, this signal is driven high
by the OPB pcore right after reset.

The FIFO is reset when the MMCM is not yet locked or when ctrl_reset is driven high by
the OPB pcore. This is controllable via software.