Fix biplex_core's determination of delay type
The biplex_core now determiens whether to use BRAMs or slices for the delays by
checking the delay length against the threshold length.  Previously it had
included the input bit width into the calculation, but this led to using more
BRAMs than desired.

Also added a check to ensure that BRAMs are not used if the delay length is not
greater than the BRAM latency.  This will prevent warnings from
fft_stage_n_init iabout overriding the use of BRAMs.