[ctrl_opb_attach.v] Removed combinational process
After noticing that occasionally after writing an arbiter override
into the ddr3_ctrl (0x81 into offset 8), we would read the software
address register (ddr3_ctrl offset 0) and see this same 0x81.  This
should not be happening. The override and the software address are
actually the only places that can be written, so this suggested
funny timing in the state machine such that these writes/reads were
crossing into the incorrect address space.  Simulation, however,
indicated no errors in timing.  Rurik noticed the always @(*) process
and suggested that piece can cause funny thigns to happen, so I
rewrote the ctrl_opb_attach.v code incorporating that logic into the
main process.  The code simulated and produced identical timing as
the previous version, but when implemented in the FPGA we no longer
saw this mistaken writing to or reading from the wrong address.