[mem_opb_attach.v] added states to OPB state machine, added probe for max_clocks
To follow OPB documentation and timing more clearly, I have added
"Select" and "transfer" states to the OPB state machine.  This has
not successfully fixed the bugs, but it makes the timing simulate
in a way that agrees with the documentation.  Also prompting this
commit is adding a probe for "max clocks" that is, the maximum
number of clock cycles to complete a single OPB action.  A fear
is that the timeout suppress signal may not be routed, and with
random latency in the ddr3 we may enter a state that aborts in
the middle of a read due to a timeout and the code is not handlign
that properly.