reorder modification
The logic to generate the sequence of addresses has been modified
so that only two buffers are required for orders greater than 2,
one contains the current sequence of addresses and the other the
map to modify this. This reduces the amount of lookup buffers required
as well as the size of the multiplexer required.