[mem_opb_attach.v] no longer hangs after sequential reads
Several adjustments here.  The first is the addition of counters
and reset counter signals added into mem_opb_attach.v.  This allowed
us information about latency and in which state of the state
machines the code was getting stuck.

After examination of these numbers, it was clear that OPB was
issuing a read, but the ddr3 state machine was not exiting
the IDLE state.  The info showed that toutsup must have been
working as expected as the OPB held in the read state for 1001
clocks (timeout occurs on 16 clocks if toutsup not held).  However,
this showed us that there is perhaps some overall timeout that
deselects the opb device after 1000(ish) clocks.

Inspecting the mem_opb_attach ddr3 state machine we saw that to
initialize a read, the ddr3 must see a "rising edge" determined by
comparing two parameters:  a read enable derived from the opb clock
and the previous read_enable sampled at the ddr3 clock.  by sampling
the opb read_en signal in the ddr3 state machine for 2 clocks, we
now compare those to find a positive edge.  Later in the ddr3 state
machine, there was also a comparison to the opb timed read_en signal
which was replaced by the ddr3 read_en signal, read_en_z.